New tool: vcd-prune
This adds a new tool, `vcd-prune`, which can be used to pull specific
modules out of a VCD file. This is useful if you're simulating a big
design (rocket-chip), but only care about the RoCC. This, for most
usage cases, dramatically reduces the filesize of the VCD file and
speeds up the start time of a waveform viewer dramatically.